refactor DNS examples
