+++ examples/dnsblock/dnsblock.lua
@@ -7,58 +7,16 @@
 
 local xt = require("xtable")
 local nf = require("netfilter")
-local linux = require("linux")
-local string = require("string")
+local common = require("examples.dnsblock.common")
 local family = nf.family
 
-local udp = 0x11
-local dns = 0x35
-
-local blacklist = {
-	"github.com",
-	"gitlab.com",
-}
-
 local function nop() end
-
-local function check_blacklist(name)
-	for _, v in ipairs(blacklist) do
-		if string.find(name, v) ~= nil then
-			return true
-		end
-	end
-	return false
-end
-
-local function get_domainname(skb, off)
-	local name = ""
-	local i = 0
-
-	repeat
-		local d = skb:getbyte(off + i)
-		i = i + 1
-	until d == 0
-
-	return name .. skb:getstring(off, i)
-end
 
 local function dnsblock_mt(skb, par)
 	local thoff = par.thoff
 	local proto = skb:getuint8(9)
 
-	if proto == udp then
-		local dstport = linux.ntoh16(skb:getuint16(thoff + 2))
-		if dstport == dns then
-			local qoff = thoff + 20
-			local name = get_domainname(skb, qoff)
-			if check_blacklist(name) then
-				print("DNS query for " .. name .. " blocked\n")
-				return true
-			end
-		end
-	end
-
-	return false
+	return common.hook(skb, thoff, proto)
 end
 
 xt.match{
